+++ libvast/src/format/syslog.cpp
@@ -105,7 +105,6 @@ const char* reader::name() const {
 
 caf::error
 reader::read_impl(size_t max_events, size_t max_slice_size, consumer& f) {
-  bool timeout = false;
   table_slice_builder_ptr bptr = nullptr;
   auto next_line = [&, start = std::chrono::steady_clock::now()] {
     auto remaining = start + read_timeout_ - std::chrono::steady_clock::now();
@@ -117,7 +116,8 @@ reader::read_impl(size_t max_events, size_t max_slice_size, consumer& f) {
     }
     return lines_->next_timeout(remaining);
   };
-  for (size_t produced = 0; produced < max_events; timeout = next_line()) {
+  for (size_t produced = 0; produced < max_events;) {
+    auto timeout = next_line();
     // We must check not only for a timeout but also whether any events were
     // produced to work around CAF's assumption that sources are always able to
     // generate events. Once `caf::stream_source` can handle empty batches