+++ libvast/src/system/pivoter.cpp
@@ -52,14 +52,11 @@ common_field(const pivoter_state& st, const record_type& indicator) {
   // updated type registry is available to feed the algorithm above.
   std::string edge;
   VAST_TRACE(st.self, VAST_ARG(st.target), VAST_ARG(indicator.name()));
-  if (detail::starts_with(st.target, "zeek")) {
-    if (detail::starts_with(indicator.name(), "zeek"))
-      edge = "uid";
-    else
-      edge = "community_id";
-  } else {
+  if (detail::starts_with(st.target, "zeek")
+      && detail::starts_with(indicator.name(), "zeek"))
+    edge = "uid";
+  else
     edge = "community_id";
-  }
   for (auto& i : indicator.fields) {
     if (i.name == edge) {
       st.cache.insert({indicator, i});