+++ libvast/vast/settings.hpp
@@ -13,22 +13,12 @@
 
 #pragma once
 
-#include "vast/aliases.hpp"
-#include "vast/fwd.hpp"
+#include <caf/settings.hpp>
 
-namespace vast::system {
+namespace vast {
 
-/// Spawns an INDEXER actor.
-/// @param parent The parent actor.
-/// @param filename File for persistent state.
-/// @param column_type The type of the indexed field.
-/// @param index_opts Runtime options to parameterize the value index.
-/// @param index A handle to the index actor.
-/// @param partition_id The partition ID that this INDEXER belongs to.
-/// @param fqn The fully qualified name of the indexed field.
-/// @returns the new INDEXER actor.
-caf::actor spawn_indexer(caf::local_actor* parent, path filename,
-                         type column_type, caf::settings index_opts,
-                         caf::actor index, uuid partition_id, std::string fqn);
+/// Merge settings of `src` into `dst`, overwriting existing values
+/// from `dst` if necessary.
+void merge_settings(const caf::settings& src, caf::settings& dst);
 
-} // namespace vast::system
+} // namespace vast