+++ libvast_test/src/dummy_index.cpp
@@ -19,7 +19,6 @@
 #include "vast/system/instrumentation.hpp"
 
 using void_fun = std::function<void()>;
-using atomic_measurement = vast::system::atomic_measurement;
 
 CAF_ALLOW_UNSAFE_MESSAGE_TYPE(void_fun)
 
@@ -38,9 +37,8 @@ dummy_indexer(caf::stateful_actor<dummy_index::dummy_indexer_state>*) {
   }};
 }
 
-caf::actor
-spawn_dummy_indexer(caf::local_actor* self, path, type, caf::settings,
-                    caf::actor, uuid, atomic_measurement*) {
+caf::actor spawn_dummy_indexer(caf::local_actor* self, path, type,
+                               caf::settings, caf::actor, uuid, std::string) {
   return self->spawn(dummy_indexer);
 }