+++ CHANGELOG.md
@@ -9,7 +9,13 @@ Every entry has a category for which we use the following visual abbreviations:
 - ⚠️ change
 - 🐞 bugfix
 
-<!-- ## Unreleased -->
+## Unreleased
+
+- 🐞 The PCAP reader now correctly shows the amount of generated events.
+  [#954](https://github.com/tenzir/vast/pull/954)
+
+- ⚠️ A type definition for the stats event was added to the suricata schema
+  file. [#954](https://github.com/tenzir/vast/pull/954)
 
 ## [2020.06.25]
 
@@ -23,6 +29,10 @@ Every entry has a category for which we use the following visual abbreviations:
   the default when built with Apache Arrow support for a while now, and the new
   name more accurately reflects what it is doing.
   [#948](https://github.com/tenzir/vast/pull/948)
+
+- 🎁 The output format for the `explore` and `pivot` commands can now be set
+  using the `explore.format` and `pivot.format` options respectively. Both
+  default to JSON. [#921](https://github.com/tenzir/vast/921)
 
 - 🎁 The meta index now uses Bloom filters for equality queries involving IP
   addresses. This especially accelerates queries where the user wants to know